+++ src/main/java/request/Request.java
@@ -1,4 +1,4 @@
-package http;
+package request;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
